requirements.txt
streamlit==1.32.0
pymongo==4.6.2
beautifulsoup4==4.12.3
requests==2.31.0
plotly==5.20.0
pydantic==2.6.4
python-dotenv==1.0.1
pandas==2.2.0
flask==3.0.0
email-validator==2.1.0